requirements.txt
requests
python-dotenv
types-requests
pytest
pytest-cov 